# Performance improvements LiveTradingResultHandler